Updated MMC1 OriginalGFX.asm for 3.0
Updated the starting address of the animated tile frames which had been shifted a few bytes between 2.0 and 3.0. This was causing a crash upon loading if this patch was used.

Added code to return Ganon's palette after he recovers from his vulnerable state to the original palette, as 3.0 modified this and it hadn't been accounted for here.

Updated "Original NES GFX.ips".